code review feedback: update variable name, move getSystemFields to utils.ts
