+++ packages/components/src/internal/components/domainproperties/dataset/utils.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import { DomainField } from '../models';
+import { DomainField, SystemField } from '../models';
 import { selectRows } from '../../../query/selectRows';
 import { SchemaQuery } from '../../../../public/SchemaQuery';
 import { caseInsensitive } from '../../../util/utils';
@@ -45,4 +45,57 @@ export const useStudyPropertiesContext = (): StudyProperties => {
         })();
     }, []);
     return state;
+};
+
+export const getDatasetSystemFields = (studyProperties: StudyProperties): SystemField[] => {
+    const isVisitBased = studyProperties.TimepointType === VISIT_TIMEPOINT_TYPE;
+
+    const systemFields = [
+        {
+            Name: studyProperties.SubjectColumnName,
+            Label: studyProperties.SubjectColumnName,
+            DataType: 'Text',
+            Required: true,
+            Description: 'Subject identifier',
+            Disableble: false,
+        },
+        {
+            Name: 'SequenceNum',
+            Label: 'Sequence Num',
+            DataType: 'Decimal (floating point)',
+            Required: isVisitBased,
+            Description: '',
+            Disableble: false,
+        },
+    ];
+
+    if (!isVisitBased) {
+        systemFields.push({
+            Name: 'date',
+            Label: 'Date',
+            DataType: 'DateTime',
+            Required: true,
+            Description: 'The day of the visit. Primarily used in date-based studies.',
+            Disableble: false,
+        });
+        systemFields.push({
+            Name: 'Day',
+            Label: 'Day',
+            DataType: 'Integer',
+            Required: false,
+            Description: 'The day of the visit. Primarily used in date-based studies.',
+            Disableble: false,
+        });
+    }
+
+    systemFields.push({
+        Name: 'DatasetId',
+        Label: 'Dataset Id',
+        DataType: 'Integer',
+        Required: true,
+        Description: '',
+        Disableble: false,
+    });
+
+    return systemFields;
 };
